# backend/src/main/java/randomTeamFC27/dto/league/LeagueCreateRequest.java
package randomTeamFC27.dto.league;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LeagueCreateRequest(@NotBlank @Size(max = 120) String name) {
}