# shared/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC27 Random Team & League Manager API",
    "description": "Backend for the FC27 random team draw and friend-league tracker",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/leagues": {
      "get": {
        "tags": [
          "Leagues"
        ],
        "summary": "List leagues owned by the current user — requires auth",
        "operationId": "listMyLeagues",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeagueResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Leagues"
        ],
        "summary": "Create a league owned by the current user — requires auth",
        "operationId": "createLeague",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/leagues/{id}/reset": {
      "post": {
        "tags": [
          "Leagues"
        ],
        "summary": "Close the league's current season and open a new one — current standings/matches go back to zero, but the closed season's scores/standings stay browsable under GET /api/leagues/{id}/seasons (\"Geçmiş Ligler\"); the league and its player roster are kept — owner only",
        "operationId": "resetLeague",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/league-types": {
      "get": {
        "tags": [
          "League Types"
        ],
        "summary": "List all league types, alphabetically — public, used to populate the team form's dropdown",
        "operationId": "listLeagueTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeagueTypeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "League Types"
        ],
        "summary": "Create a new league type — ADMIN only. 400 if the name already exists (case-insensitive)",
        "operationId": "createLeagueType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeagueTypeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueTypeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Rotate a refresh token (body or cookie) for a new access+refresh pair — public",
        "operationId": "refresh",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/otp/sms/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify the texted code and sign in (or register) — public",
        "operationId": "verifySmsOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsOtpVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/otp/sms/send": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Send a 6-digit SMS login code — public, rate-limited",
        "operationId": "sendSmsOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsOtpSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/otp/email/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify the emailed code and sign in (or register) — public",
        "operationId": "verifyEmailOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOtpVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/otp/email/send": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Send a 6-digit email login code — public, rate-limited",
        "operationId": "sendEmailOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOtpSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Revoke the caller's refresh token and clear the cookie — requires a valid access token",
        "operationId": "logout",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/google": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign in (or register) with a Google ID token — public",
        "operationId": "loginWithGoogle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LeagueCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          }
        }
      },
      "LeagueResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LeagueTypeCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          }
        }
      },
      "LeagueTypeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummaryDto"
          }
        }
      },
      "UserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        }
      },
      "SmsOtpVerifyRequest": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\+?[0-9]{7,15}"
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\d{6}"
          },
          "rememberMe": {
            "type": "boolean"
          }
        }
      },
      "SmsOtpSendRequest": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\+?[0-9]{7,15}"
          }
        }
      },
      "EmailOtpVerifyRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\d{6}"
          },
          "rememberMe": {
            "type": "boolean"
          }
        }
      },
      "EmailOtpSendRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "GoogleLoginRequest": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "minLength": 1
          },
          "rememberMe": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}